Add files via upload
ABSTRACT


Consumer protection is a consequence of legal force in meeting the interests of consumers where consumers are the basis of complete economic activity in terms of achieving the gool of produktivity planner review (improving the quality of work and products/goods/services), this requires a balance between the quality/quality of services and the balance of the level of conument satisfaction. So producers / services must further improve the quality of goods in a fairness (honest) profit informing the quality of goods, namely whether the goods / services have etiquette (information) such as Halal labels, Inferior Goods Labels (low quality), Essential Goods labels and luxury goods labels . The importance of information is the strength of a good legal system so that consumers and producers / services alike increase their trust in the changes of the demands of this era of globalization. The force of law in an information is a means of consumer protection in an agreement that is mutually binding between rights and obligations.
In the theory of opportunity (opportunity) producers see many opportunities to achieve mutual benefits to compete with each other to manage home industries with quality adapted to the Goods /Services Quality Label.  as stipulated in Article 7 of the Consumer Protection Law on its holiness as a business actor, developers / developers participate in establishing honest economic power regulated in the Housing Law and obey  the rules of the housing association, as well as developers in this case Apartments are able to annul their business activities in order to prevent disputes or dispute.
Consumer protection law as the basis for the peace settlement under the hands of the Problem "Settlement of Apartment Arrears Between Mrs. MP and PT. X a subsidiary of PT. Y Connected With The Consumer Protection Act", is the title the author gives in this Thesis. Through negotiations the parties to the dispute may undertake a process of re-exploring the rights and obligations of the parties to the dispute in arrears with a situation that is equally favorable, by releasing or granting leeway to certain rights based on the principle of reciprocity. Where the agreement that has been reached is stated in writing to be signed and executed by the parties.